Reorganized originate.js into origination.js.
