+++ lib/core/stateMachine.js
@@ -5,11 +5,12 @@ var util = require('util');
 var Q = require('q');
 
 var getOrCreateBridge = require('./helpers/getOrCreateBridge.js');
-var originate = require('./helpers/originate.js');
+var origination = require('./helpers/origination.js');
 
 function create(client, data, channel) {
   var bridge;
   var participants = {};
+  var originator;
 
   return {
     states: {
@@ -68,12 +69,14 @@ function create(client, data, channel) {
       bridge.on('ChannelEnteredBridge', this.onChannelEnteringBridge);
       bridge.on('ChannelLeftBridge', this.onChannelExitingBridge);
 
-      originate.preOriginate.call(this, {
+      this.originator = origination.call(this, {
         client: client,
         channel: channel,
         bridge: bridge,
         data: data
       });
+
+      this.originator.preOriginate();
 
       channel.once('ChannelHangupRequest', this.onChannelHangup);
     },
@@ -94,6 +97,8 @@ function create(client, data, channel) {
       var deviceState = Q.denodeify(
         client.deviceStates.update.bind(client)
       );
+
+      this.currentState = state;
 
       return deviceState({
         deviceName: util.format('Stasis:%s', data.extension.name),
@@ -152,9 +157,13 @@ function create(client, data, channel) {
 
           this.allowDtmf = false;
 
-          originate.originate(util.format('SIP/%s@%s',
-                                          this.dialString,
-                                          data.extension.trunks[0]));
+          this.originator.originate(
+            util.format(
+              'SIP/%s@%s',
+              this.dialString,
+              data.extension.trunks[0]
+            )
+          );
 
           break;
